+++ backend/canisters/local_user_index/api/src/queries/notifications_v2.rs
@@ -3,7 +3,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use types::{IndexedEvent, NotificationBotDetails, NotificationEnvelope, SubscriptionInfo, TimestampMillis, UserId};
+use types::{IndexedEvent, NotificationEnvelope, SubscriptionInfo, TimestampMillis, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -19,6 +19,6 @@ pub enum Response {
 pub struct SuccessResult {
     pub notifications: Vec<IndexedEvent<NotificationEnvelope>>,
     pub subscriptions: HashMap<UserId, Vec<SubscriptionInfo>>,
-    pub bots: HashMap<UserId, NotificationBotDetails>,
+    pub bot_endpoints: HashMap<UserId, String>,
     pub timestamp: TimestampMillis,
 }