+++ backend/canisters/local_user_index/impl/src/updates/c2c_notify_user_index_events.rs
@@ -12,10 +12,7 @@ use stable_memory_map::StableMemoryMap;
 use std::cell::LazyCell;
 use std::cmp::min;
 use tracing::info;
-use types::{
-    BotEvent, BotEventWrapper, BotLifecycleEvent, BotNotificationEnvelope, BotRegisteredEvent, NotificationEnvelope,
-    TimestampMillis, c2c_uninstall_bot,
-};
+use types::{BotEvent, BotLifecycleEvent, BotNotification, BotRegisteredEvent, TimestampMillis, c2c_uninstall_bot};
 use user_canister::{
     DiamondMembershipPaymentReceived, DisplayNameChanged, ExternalAchievementAwarded, OpenChatBotMessageV2,
     PhoneNumberConfirmed, ReferredUserRegistered, StorageUpgraded, UserJoinedCommunityOrChannel, UserJoinedGroup,
@@ -110,20 +107,17 @@ fn handle_event<F: FnOnce() -> TimestampMillis>(
             let this_canister_id = state.env.canister_id();
             if ev.notification_canister == this_canister_id {
                 // This local_user_index canister has been nominated to notify the bot
-                state
-                    .data
-                    .notifications
-                    .add(NotificationEnvelope::Bot(BotNotificationEnvelope {
-                        event: BotEventWrapper {
-                            api_gateway: this_canister_id,
-                            event: BotEvent::Lifecycle(BotLifecycleEvent::Registered(BotRegisteredEvent {
-                                bot_id: ev.bot_id,
-                                bot_name: ev.name,
-                            })),
-                        },
+                state.data.push_bot_notification(
+                    BotNotification {
+                        event: BotEvent::Lifecycle(BotLifecycleEvent::Registered(BotRegisteredEvent {
+                            bot_id: ev.bot_id,
+                            bot_name: ev.name,
+                        })),
                         recipients: vec![ev.bot_id],
-                        timestamp: **now,
-                    }));
+                    },
+                    this_canister_id,
+                    **now,
+                );
             }
         }
         UserIndexEvent::BotPublished(ev) => {