+++ backend/notification_pusher/aws/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Serialize bot notifications in LUI not pusher ([#8169](https://github.com/open-chat-labs/open-chat/pull/8169))
+
+## [[2.0.1789](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1789-notifications_pusher)] - 2025-06-13
+
 ### Added
 
 - Updated the way configuration is loaded for the pusher by using the `envconf` crate ([8145](https://github.com/open-chat-labs/open-chat/pull/8145))